+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultStainCategoryService.java
@@ -10,6 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import uk.ac.bbsrc.tgac.miso.core.data.StainCategory;
 import uk.ac.bbsrc.tgac.miso.core.store.DeletionStore;
+import uk.ac.bbsrc.tgac.miso.core.util.Pluralizer;
 import uk.ac.bbsrc.tgac.miso.persistence.StainCategoryDao;
 import uk.ac.bbsrc.tgac.miso.service.StainCategoryService;
 import uk.ac.bbsrc.tgac.miso.service.exception.ValidationError;
@@ -88,8 +89,7 @@ public class DefaultStainCategoryService implements StainCategoryService {
     ValidationResult result = new ValidationResult();
     long usage = stainCategoryDao.getUsage(object);
     if (usage > 0L) {
-      result.addError(new ValidationError("Stain category '" + object.getName() + "' is used by " + usage + " stain"
-          + (usage > 1L ? "s" : "")));
+      result.addError(ValidationError.forDeletionUsage(object, usage, Pluralizer.stains(usage)));
     }
     return result;
   }